pyproject.toml
[tool.poetry]
name = "pnm-saltokuppers"
version = "0.1.0"
description = ""
authors = ["Felipe Sodré M. Barros"]
readme = "README.md"
packages = [{include = "pnm_saltokuppers"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.23"
geoalchemy2 = "^0.14.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
ruff = "^0.1.6"
isort = "^5.12.0"
black = "^23.11.0"
sqlalchemy-utils = "^0.41.1"
geopandas = "^0.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
